(PDK-1364) Rename templatedir.rb
The templatedir.rb file has the TemplateDir module in there however, the
filename should really be template_dir.rb according to ruby naming standards.
This commit renames the file and updates any locations which use it.
