+++ lib/pdk/module/template_dir.rb
